+++ R/motifs.R
@@ -6,6 +6,7 @@ NULL
 #' @method AddMotifs default
 #' @concept motifs
 #' @importFrom methods slot
+#' @importFrom GenomeInfoDb seqlevels seqnames
 #' @export
 AddMotifs.default <- function(
   object,
@@ -28,6 +29,14 @@ AddMotifs.default <- function(
   if (verbose) {
     message("Building motif matrix")
   }
+  # genome can be string
+  if (is.character(x = genome)) {
+    if (!requireNamespace("BSgenome", quietly = TRUE)) {
+      stop("Please install BSgenome.
+             https://www.bioconductor.org/packages/BSgenome/")
+    }
+    genome <- BSgenome::getBSgenome(genome = genome)
+  }
   motif.matrix <- CreateMotifMatrix(
     features = object,
     pwm = pfm,
@@ -37,9 +46,14 @@ AddMotifs.default <- function(
   if (verbose) {
     message("Finding motif positions")
   }
+  
+  # for positions, a list of granges is returned
+  # each element of list is a PFM name
+  # each entry in granges is the position within a feature that matches motif
+  obj_keep <- as.character(seqnames(x = object)) %in% seqlevels(x = genome)
   motif.positions <- motifmatchr::matchMotifs(
     pwms = pfm,
-    subject = object,
+    subject = object[obj_keep],
     out = 'positions',
     genome = genome
   )